vsyscall arm - use uint32_t typedefs for tv_sec/nsec

+++ arch/arm/kernel/update_vsyscall_arm.c
@@ -43,8 +43,8 @@ struct kernel_gtod_t {
 	u64  mask;
 	u32  mult;
 	u32  shift;
-	unsigned long tv_sec;
-	unsigned long tv_nsec;
+	uint32_t tv_sec;
+	uint32_t tv_nsec;
 };
 
 struct kernel_tz_t {
@@ -53,8 +53,8 @@ struct kernel_tz_t {
 };
 
 struct kernel_wtm_t {
-	unsigned long  tv_sec;
-	unsigned long  tv_nsec;
+	uint32_t  tv_sec;
+	uint32_t  tv_nsec;
 };
 
 /*
